% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomender.R
\name{biomender}
\alias{biomender}
\title{Predict using the biomedical journal recommendation model}
\usage{
biomender(x, k = 10)
}
\arguments{
\item{x}{a vector containing abstract text}

\item{k}{integer with the number of predictions (journals) to make. Defaults to 10}
}
\value{
a matrix in formate m (m abstract text) rows and k (top k suggested journals) columns
}
\description{
Predict using the biomedical journal recommendation model
}
\examples{
\dontrun{
library(biomenderR)

x <- "In this paper, we proposed biomenderR, an R-based biomedical journal \
recommendation framework using abstract text embedding approach. biomenderR \
runs in user-friendly R environment and is convenient in application \
without deep learning knowledge.The core model was trained based on \
a large dataset containing 6,240,122 abstract texts from 1352 journals \
extracted from the PubMed database. Our framework is lightweight costing \
about 100 MB of hard-disk space while maintaining equivalent performance \
to state-of-the-art methods. Moreover, we innovatively applied this framework \
to provide an in-depth analysis on the characteristics and associations of \
research fields covered by more than 1000 biomedical journals worldwide."

y <- biomender(x)
}
}
